+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug where some connectivity matrices wouldn't generate if anatomical and functional outputs were in different resolutions.
 - Handling of `3dECM` outputs for AFNI ≥ 21.1.1.
 - Fixed a bug where sparsity thresholds were not being scaled for network centrality.
+- Fixed a bug where `calculate_motion_first` would not calculate motion at all.
 
 ### Changed